requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
cassandra-driver==3.29.1
cassio==0.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.48
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.20.0
orjson==3.10.1
packaging==23.2
pydantic==2.7.0
pydantic-core==2.18.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
uvicorn==0.29.0
yarl==1.9.4
